requirements.txt
requests
beautifulsoup4
python-dotenv
pymongo
pandas
selenium
nltk
mongomock
pytest
kafka-python
fastapi
uvicorn
pandas
transformers
sentencepiece